# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Local server"
        },
        {
            "url": "https://foodie-app.azurewebsites.net",
            "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "foodie"
            }
        },
...
    },
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "security": [],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "Username",
                                        "example": "test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "description": "Email",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "Password",
                                        "example": "test"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "username": {
                                            "type": "string",
                                            "description": "Username",
                                            "example": "test"
                                        },
                                        "password": {
                                            "type": "string",
                                            "description": "Password",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        },
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string",
                                    "description": "Set-Cookie"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/authenticate": {
            "post": {
                "security": [],
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "description": "Login a user",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "Username",
                                        "example": "test"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "Password",
                                        "example": "test"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "username": {
                                            "type": "string",
                                            "description": "Username",
                                            "example": "test"
                                        },
                                        "password": {
                                            "type": "string",
                                            "description": "Password",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        },
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string",
                                    "description": "Set-Cookie"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "delete": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout a user",
                "description": "Logout a user",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Message",
                                            "example": "Logout successful"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/check-session": {
            "get": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Check if user is logged in",
                "description": "Check if user is logged in",
                "operationId": "checkSession",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "username": {
                                            "type": "string",
                                            "description": "Username",
                                            "example": "test"
                                        },
                                        "password": {
                                            "type": "string",
                                            "description": "Password",
                                            "example": "test"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}